moved profile sections to individual components
